+++ src/2_setup/setup_mlx.c
@@ -6,7 +6,7 @@
 /*   By: freddy <freddy@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/05/28 08:41:35 by freddy            #+#    #+#             */
-/*   Updated: 2024/06/05 18:53:23 by freddy           ###   ########.fr       */
+/*   Updated: 2024/06/11 15:47:08 by freddy           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -56,5 +56,4 @@ void	setup_mlx(void)
 	setup_player();
 	mlx_loop_hook(data->mlx, loop_hook, NULL);
 	mlx_key_hook(data->mlx, key_hook, NULL);
-	mlx_scroll_hook(data->mlx, scroll_hook, NULL);
 }