+++ src/3_game_loop/a_game_logic/hooks/loop_hook.c
@@ -6,15 +6,35 @@
 /*   By: freddy <freddy@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/05/28 10:23:28 by freddy            #+#    #+#             */
-/*   Updated: 2024/06/11 14:13:59 by freddy           ###   ########.fr       */
+/*   Updated: 2024/06/11 16:14:46 by freddy           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../../../../include/cub3d.h"
 
+static int	get_frame(double time)
+{
+	return (int)((time - floor(time)) * FPS);
+}
+
 void	loop_hook(void *param)
 {
-	render_game_scene();
-	// printf("POS: [%f, %f] - ROT: [%f, %f]: %d\n", player()->transform.pos.x, player()->transform.pos.y, player()->transform.rot.x, player()->transform.rot.y, *game()->input_data->map[(int)player()->transform.pos.y][(int)player()->transform.pos.x]);
+	double	curr_time;
+	double	prev_time;
+	int		curr_frame;
+	int		prev_frame;
+
 	(void) param;
+	curr_time = mlx_get_time();
+	prev_time = game()->prev_time;
+	curr_frame = get_frame(curr_time);
+	prev_frame = get_frame(prev_time);
+	if (curr_frame > prev_frame + 1)
+		logger(LOGGER_WARNING, "skipped frame");
+	if ((int)curr_time != (int)prev_time || curr_frame != prev_frame)
+	{
+		render_game_scene();
+		handle_mouse_mv();
+		game()->prev_time = curr_time;
+	}
 }