+++ src/3_game_loop/b_rendering/render.c
@@ -6,7 +6,7 @@
 /*   By: freddy <freddy@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/05/30 14:36:45 by freddy            #+#    #+#             */
-/*   Updated: 2024/06/11 14:33:40 by freddy           ###   ########.fr       */
+/*   Updated: 2024/06/11 15:49:18 by freddy           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -18,6 +18,6 @@ void	render_game_scene(void)
 	game()->game_scene = mlx_new_image(game()->mlx, game()->mlx->width, game()->mlx->height);
 	do_wall_operations();
 	mlx_image_to_window(game()->mlx, game()->game_scene, 0, 0);
-	printf("\033[2J");
+	// printf("\033[2J");
 	print_map(game()->input_data->map, ".01");
 }